修改了dataextractor的代码，在本地集群能跑的动了，青云上估计还是悬。看看吧。

+++ src/main/scala/org/brave/spark/caseclass/Ratings.scala
@@ -1,4 +1,4 @@
 package org.brave.spark.caseclass
 
-case class Ratings(userId: Int, movieId: Int, rating: Float, timestamp: Double)
+case class Ratings(userId: Int, movieId: Int, rating: Float, timestamp: Int)
 //配合ETL构建表
